+++ lib/RGBLamp/RGBLamp.h
@@ -24,8 +24,8 @@ class RGBLamp
     int           _current_color[3];
     int           _fade_from[3];
     int           _fade_to[3];
-    float         _fade_start;
-    float         _fade_end;
+    unsigned long _fade_start;
+    unsigned long _fade_end;
     bool _fade;
 };
